refactor: move implementation inside anymap crate into stdx crate
