--- crates/anymap/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "anymap"
-version = "0.0.0"
-description = "This crate is a port of only the necessary features from https://github.com/chris-morgan/anymap for use within rust-analyzer. Copyright © 2014–2022 Chris Morgan. COPYING: https://github.com/chris-morgan/anymap/blob/master/COPYING"
-
-authors.workspace = true
-edition.workspace = true
-license.workspace = true
-rust-version.workspace = true
-
-[package.metadata.docs.rs]
-all-features = true
-
-[features]
-default = ["std"]
-std = []
-
-[dependencies]
-# The hashbrown feature, disabled by default, is exposed under different stability guarantees than the usual SemVer ones: by preference the version range will only be extended, but it may be shrunk in a MINOR release. See README.md.
-hashbrown = { version = "0.14.0", optional = true }
