+++ crates/stdx/Cargo.toml
@@ -17,6 +17,7 @@ backtrace = { version = "0.3.67", optional = true }
 always-assert = { version = "0.1.2", features = ["log"] }
 jod-thread = "0.1.2"
 crossbeam-channel = "0.5.5"
+hashbrown = { version = "0.14.0" }
 # Think twice before adding anything here
 
 [target.'cfg(windows)'.dependencies]
